AddLocationFragment
Correctly tests if the name and point chosen exists within the database already. If it does a toast is triggered to notify the user if the name, point, or both already exists. Need to correct navigation and prevent it from going back to main activity only if a new location is created.